Dialogs erstellt bevor man löscht oder updatet

+++ CYR/ViewModel/ClientViewModel.cs
@@ -1,10 +1,13 @@
-﻿using System.Collections.ObjectModel;
-using CommunityToolkit.Mvvm.ComponentModel;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using CYR.Clients;
 using CYR.Clients.ViewModels;
+using CYR.Dialog;
 using CYR.Invoice.InvoiceViewModels;
 using CYR.Services;
+using System.Collections.ObjectModel;
+using System.Linq.Expressions;
+using System.Windows;
 
 namespace CYR.ViewModel;
 
@@ -12,12 +15,17 @@ public partial class ClientViewModel : ObservableObject, IParameterReceiver
 {
     private readonly IRetrieveClients _retrieveClients;
     private readonly IClientRepository _clientRepository;
-    public ClientViewModel(IRetrieveClients retrieveClients, INavigationService navigationService, IClientRepository clientRepository)
+    private readonly IDialogService _dialogService;
+
+    private string? _dialogResponse;
+
+    public ClientViewModel(IRetrieveClients retrieveClients, INavigationService navigationService, IClientRepository clientRepository, IDialogService dialogService)
     {
         _retrieveClients = retrieveClients;
         Navigation = navigationService;
         Initialize();
         _clientRepository = clientRepository;
+        _dialogService = dialogService;
     }
 
     private async void Initialize()
@@ -65,6 +73,12 @@ public partial class ClientViewModel : ObservableObject, IParameterReceiver
             {
                 try
                 {
+                    ShowNotificationDialog("Kunde löschen", $"Möchten Sie wirklich den Kunden löschen? Das würde bedeuten, dass alle Rechnungen und Daten im Zusammenhang mit diesen Kudnen auch gelsöcht werden!",
+                    "Nein", "User", Visibility.Visible, "Ja");
+                    if (_dialogResponse != "True")
+                    {
+                        return;
+                    }
                     var c = await _clientRepository.DeleteAsync(client);
                 }
                 catch (Exception)
@@ -91,5 +105,25 @@ public partial class ClientViewModel : ObservableObject, IParameterReceiver
         var clientToUpdate = Clients.First(c => c.IsSelected);
         if (clientToUpdate is null) return;
         Navigation.NavigateTo<UpdateClientViewModel>(clientToUpdate);
+    }
+    private void ShowNotificationDialog(string title,
+            string message,
+            string cancelButtonText,
+            string icon,
+            Visibility okButtonVisibility, string okButtonText)
+    {
+        _dialogService.ShowDialog(result =>
+        {
+            _dialogResponse = result;
+        },
+        new Dictionary<Expression<Func<NotificationViewModel, object>>, object>
+        {
+                { vm => vm.Title, title },
+                { vm => vm.Message,  message},
+                { vm => vm.CancelButtonText, cancelButtonText },
+                { vm => vm.Icon,icon },
+                { vm => vm.OkButtonText, okButtonText },
+                { vm => vm.IsOkVisible, okButtonVisibility}
+        });
     }
 }
